requirements.txt
django==5.1.3
djangorestframework==3.15.2

drf-yasg==1.21.8
python-dotenv==1.0.1

djangorestframework-simplejwt==5.3.1
PyJWT==2.10.1

psycopg2-binary==2.9.10